Separate Recruiter/User signups
Signing up as a user adds you to the user table. Signing up as a recruiter adds you to the user table and then adds you to the recruiter table.

Only userId is stored in the active session. Next step is to be able to differentiate users from userrecruiters in the session.